Enhance directory checks & refactoring
